+++ src/django_bird/__init__.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
+from pluggy import HookimplMarker
+
 __version__ = "0.14.2"
+
+hookimpl = HookimplMarker("django_bird")