+++ src/django_bird/plugins/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .manager import pm
+
+__all__ = [
+    "pm",
+]