+++ src/django_bird/plugins/hookspecs.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from pluggy import HookspecMarker
+
+hookspec = HookspecMarker("django_bird")